+++ components/box/src/actions/deleteObject.ts
@@ -2,6 +2,7 @@ import { action, util } from "@prismatic-io/spectral";
 import { createAuthorizedClient } from "../client";
 import { path, connectionInput } from "../inputs";
 import { getPathEntries } from "../utils";
+import { deleteObjectExamplePayload } from "../examplePayloads";
 export const deleteObject = action({
   display: {
     label: "Delete Object",
@@ -11,17 +12,17 @@ export const deleteObject = action({
     const client = createAuthorizedClient({ boxConnection });
     const pathEntries = await getPathEntries(client, util.types.toString(path));
     const { id, type, name } = pathEntries.slice(-1)[0];
-    let result: unknown;
     if (type === "folder") {
-      result = await client.folders.delete(id);
+      await client.folders.deleteFolderById(util.types.toString(id));
     } else if (type === "file") {
-      result = await client.files.delete(id);
+      await client.files.deleteFileById(util.types.toString(id));
     } else {
       throw Error(`'${name}' is neither a file nor a folder`);
     }
     return {
-      data: result,
+      data: {},
     };
   },
   inputs: { path, boxConnection: connectionInput },
+  examplePayload: deleteObjectExamplePayload,
 });